RAG_AGENT/pyproject.toml
[project]
name = "rag_agent"
version = "0.1.0"
authors = [
    {name = "seshagiri", email = "kondayapalepuseshagiri@gmail.com"}]
requires-python = ">=3.12"
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = ">=3.12"
requests = "^2.31.0"
google-cloud-storage="^2.19.0"
google-genai="^1.14.0"
gitpython="^3.1.40"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
absl-py = "^2.1.0"
cloudpickle = "^3.0.0"
deprecated = "^1.2.18"
google-cloud-aiplatform = {extras = ["adk", "agent_engines"], version = "^1.97.0"}
google-adk = "^1.0.0"
llama-index-core = {version = "^0.12.43", python = ">=3.12"}
[tool.poetry.scripts]
adk-short-bot = "rag_agent:app"
deploy-local = "deployment.local:main"
deploy-remote = "deployment.remote:main"
cleanup = "deployment.cleanup:cleanup_deployment"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

 